chore: extract table general settings to its own component
